% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabPackageCode.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Package LinReg}
\value{
Print out the coefficients and coefficient names, similar as done by the lm class.
}
\description{
Package LinReg
}
\section{Fields}{

\describe{
\item{\code{x}}{matrix.}

\item{\code{y}}{numeric.}

\item{\code{BetaHat}}{matrix.}

\item{\code{yHat}}{matrix.}

\item{\code{eHat}}{matrix.}

\item{\code{Df}}{numeric.}

\item{\code{sigmaSquare}}{matrix.}

\item{\code{varOfBetaHat}}{matrix.}

\item{\code{tOfEachcoefficientAmount}}{numeric.}

\item{\code{calculation}}{numeric.}

\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{namedvector}}{numeric.}
}}


\examples{
data(iris)
mod_object <- lm(Petal.Length~Species, data = iris)
print(mod_object)
}
